+++ packages/sdk-react/src/components/full/MeetingRoom.tsx
@@ -25,7 +25,6 @@ function MeetingRoomBase({
 	canManageParticipants = false,
 	onToggleParticipantMute,
 	onRemoveParticipant,
-	onUpdateDisplayName,
 	activeReactions = [],
 	isMuted = false,
 	isVideoEnabled = false,
@@ -423,7 +422,6 @@ function MeetingRoomBase({
 					canManageParticipants={canManageParticipants}
 					onToggleParticipantMute={onToggleParticipantMute}
 					onRemoveParticipant={onRemoveParticipant}
-					onUpdateDisplayName={onUpdateDisplayName}
 					onAddPeople={handleAddPeople}
 					chatMessages={chatMessages}
 					onSendMessage={onSendMessage}